pyproject.toml
[project]
name = "ai-sales"
version = "0.1.0"
description = ""
authors = [
    {name = "Aidar Darmesh",email = "darmesh.aidar@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "librosa (>=0.11.0,<0.12.0)",
    "soundfile (>=0.13.1,<0.14.0)",
    "dotenv (>=0.9.9,<0.10.0)",
    "google-genai (>=1.21.1,<2.0.0)",
    "fastapi (>=0.115.13,<0.116.0)",
    "uvicorn (>=0.34.3,<0.35.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
